backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
pydantic[email]==2.5.0
pytest==7.4.3
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pdfplumber==0.10.3
google-generativeai==0.3.2
tenacity==8.2.3
email-validator==2.1.0
slowapi==0.1.9
python-magic==0.4.27
redis==5.0.1
celery==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
